// aries-indeed_framework/backend/src/generic.modules/crypto/keyListArgs.enum.ts
export enum entities {
    SMA = 'SMA',
    SMAprobanden = 'SMA Probanden',
    FfE = 'FfE',
    Schweiger = 'Schweiger',
    Rheinenergie = 'Rheinenergie'
}

export enum roles {
    consumer = 'consumers',
    producer = 'producers',
}

export enum facilities {
    // SMA
    ger04600 = 'Germany|04600',
    ger04603 = 'Germany|04603',
    ger08301 = 'Germany|08301',
    ger08309 = 'Germany|08309',
    ger25917 = 'Germany|25917',
    ger59609 = 'Germany|59609',
    ger71154 = 'Germany|71154',
    ger85567 = 'Germany|85567',
    ger85656 = 'Germany|85656',
    ger91177 = 'Germany|91177',
    // SMA Probanden
    proband01 = '425710900',
    proband02 = '440743500',
    proband03 = '306603400',
    proband04 = '622216400',
    proband05 = '205858500',
    proband06 = '571190700',
    proband07 = '559421200',
    proband08 = '526103400',
    proband09 = '541360200',
    proband10 = '522700200',
    proband11 = '609687200',
    proband12 = '563550000',
    proband13 = '398940500',
    proband14 = '597847700',
    proband15 = '311377700',
    proband16 = '217379700',
    proband17 = '121969800',
    // FfE
    hq =  'HQ',
    // Schweiger
    ews1 = 'EWS1',
    ews2 = 'EWS2',
    ews3 = 'EWS3',
    ews4 = 'EWS4',
    ews5 = 'EWS5',
    ews6 = 'EWS_Gewerbe_2',
    ews7 = 'EWS_Gewerbe_1',
    // Rheinenergie
    ze01 = 'ZE01 AG06',
    ze02 = 'ZE02 AG06',
    ze03 = 'ZE03 AG06',
    ze04 = 'ZE04 AG06',
    ze05 = 'ZE05 AG06',
    ze06 = 'ZE06 AG06',
}